Add level loader
You can now load levels in the LevelLoader scene
This scene is not necessarily supposed to be the scene used in the
machine learning agent, but shows how the levels can be loaded using the
level loader script.
I created the player and the wall prefabs from the objects in the basic
scene.